SAPPHO-1: Slightly surreal solution to the indexing issue that plagued the auto-bump after approvals.
It seems that WorkflowTransitionUtil doesn't reindex an issue after transitioning it. There's a lot of additional guff on the net about this, including the mildly frightening http://forums.atlassian.com/thread.jspa?messageID=257303594 which suggests even more might be needed. One to keep a bit more of an eye on!